Simplify TypeIsGadt Interface
This commit is the first steps towards a major simplication of how
inductives are dealt internally.

Our first goal is to remove the return type of the constructors being
options. For example

```
type 'a foo =
  | X : 'a -> 'a foo
  | Y of string

```

In this case, `Y` will have type `None`, our goal is to make it have
type `'a foo`.