Vendor Merlin
This patch removes the codebase of Merlin from the repository, and add
it back as a git submodule.  We use the fork used by `ocaml-lsp', that
has the particularity to assign a public name to Merlin libraries.
Other than that, Merlin is left unchanged.  We need to adapt
coq-of-ocaml to that.  It turns out that the `assert false' reached by
coq-of-ocaml can be avoided by using `Mpipeline.with_pipeline'.